Factorize code in `backward`
