+++ src/torchjd/backward.py
@@ -4,11 +4,6 @@ from torch import Tensor
 
 from ._transform import Diagonalize, EmptyTensorDict, Init, Jac, Store, make_aggregation
 from ._transform.strategy import UnifyingStrategy
-from ._utils import (
-    _as_tensor_list,
-    _test_non_negative_optional_chunk_size,
-    _test_retain_graph_compatible_with_chunk_size,
-)
 from .aggregation import Aggregator
 
 
@@ -59,11 +54,14 @@ def backward(
         larger value results in faster differentiation, but also higher memory usage. Defaults to
         ``None``.
     """
-    _test_non_negative_optional_chunk_size(parallel_chunk_size)
+    _check_non_negative_optional_chunk_size(parallel_chunk_size)
 
     tensors = _as_tensor_list(tensors)
 
-    _test_retain_graph_compatible_with_chunk_size(tensors, retain_graph, parallel_chunk_size)
+    if len(tensors) == 0:
+        raise ValueError("`tensors` cannot be an empty iterable of `Tensor`s.")
+
+    _check_retain_graph_compatible_with_chunk_size(tensors, retain_graph, parallel_chunk_size)
 
     inputs = list(inputs)
 
@@ -85,3 +83,32 @@ def backward(
     backward_transform = store << aggregation << jac << diag << init
 
     backward_transform(EmptyTensorDict())
+
+
+def _check_non_negative_optional_chunk_size(parallel_chunk_size: int | None) -> None:
+    if not (parallel_chunk_size is None or parallel_chunk_size > 0):
+        raise ValueError(
+            "`parallel_chunk_size` should be `None` or greater than `0`. (got "
+            f"{parallel_chunk_size})"
+        )
+
+
+def _as_tensor_list(tensors: Sequence[Tensor] | Tensor) -> list[Tensor]:
+    if isinstance(tensors, Tensor):
+        output = [tensors]
+    else:
+        output = tensors
+    return output
+
+
+def _check_retain_graph_compatible_with_chunk_size(
+    tensors: list[Tensor],
+    retain_graph: bool,
+    parallel_chunk_size: int | None,
+) -> None:
+    tensors_numel = sum([tensor.numel() for tensor in tensors])
+    if parallel_chunk_size is not None and parallel_chunk_size < tensors_numel and not retain_graph:
+        raise ValueError(
+            "When using `retain_graph=False`, parameter `parallel_chunk_size` must be `None` or "
+            "large enough to compute all gradients in parallel."
+        )
